pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fluxo"
version = "0.1.0"
description = "A modern, open-source data orchestration and ELT tool."
authors = [
    { name = "Fluxo Contributors", email = "hello@fluxo.dev" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "duckdb>=0.9.0",
    "streamlit>=1.28.0",
    "streamlit-agraph>=0.0.45",
    "networkx>=3.0",
    "pydantic>=2.0.0",
    "sqlglot>=20.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
fluxo = "fluxo.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
